requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
cachetools==4.2.2
    # via google-auth
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via google-crc32c
charset-normalizer==2.0.4
    # via requests
django==3.2.7
    # via
    #   -r requirements.in
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
django-environ==0.6.0
    # via -r requirements.in
django-storages==1.11.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.8.0
    # via -r requirements.in
google-api-core[grpc]==2.0.1
    # via
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.0.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.0.0
    # via google-cloud-storage
google-cloud-secret-manager==2.7.0
    # via -r requirements.in
google-cloud-storage==1.42.1
    # via -r requirements.in
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==2.0.2
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.53.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
grpc-google-iam-v1==0.12.3
    # via google-cloud-secret-manager
grpcio==1.40.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
idna==3.2
    # via requests
libcst==0.3.20
    # via google-cloud-secret-manager
mypy-extensions==0.4.3
    # via typing-inspect
mysqlclient==2.0.3
    # via -r requirements.in
packaging==21.0
    # via google-cloud-secret-manager
proto-plus==1.19.0
    # via google-cloud-secret-manager
protobuf==3.17.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.20
    # via cffi
pyjwt==2.1.0
    # via djangorestframework-simplejwt
pyparsing==2.4.7
    # via packaging
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via libcst
requests==2.26.0
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   grpcio
    #   protobuf
sqlparse==0.4.2
    # via django
typing-extensions==3.10.0.2
    # via
    #   libcst
    #   typing-inspect
typing-inspect==0.7.1
    # via libcst
urllib3==1.26.6
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
